sir i did it
i finished almost all of my code, I believe all I have left is the data visuilisation and making it look nice. firstly I split the original score column into a score and deviser which I could use to get a percentage score, but in trying to convert them too floats I found alot of mistakes in the data that I had to fix including a missing value that wouldn't go away and I had to manually find and delete as I couldn't use ctrl + f on nothing, and it didn't count as a missing value so the nan thing didn't work. I then made a new dataframe to average all of the scores which took half an hour to try to fix as I had forgot a bracket in one of the lines. I then made a id list with all of the creation dates and dropped all duplicate columns gibing only the first creation date in the csv file for each id. I then had to make a new dataframe to put the data in which presented problems with me having to reset the index for the id_df dates to line up properly. now I have most of what I need for visualisation.